New code.
I was having problems with my laser pointer bluetooth.ino program.  It
might have been due to the android side or maybe arduino side, not sure.

This program takes in strings of values and directs them to proper
functions such as changing values of x or y axis or laser switch.